fix: изменён тип amount на java.lang.Double в NavigateToPayment event

+++ app/src/main/java/com/example/hotel_app/presentation/ui/fragments/BookingFragment.kt
@@ -197,7 +197,7 @@ class BookingFragment : Fragment(R.layout.fragment_booking) {
                             is BookingViewModel.BookingUiEvent.NavigateToPayment -> {
                                 // Переход на оплату с передачей суммы
                                 val action = BookingFragmentDirections.actionBookingFragmentToPaymentFragment(
-                                    amount = event.amount.toDouble(),
+                                    amount = event.amount,
                                     bookingId = event.booking.id,
                                     roomNumber = event.booking.roomNumber
                                 )
